Extracted DropboxConnectController to separate file, refactored mulitplatform compatibility via better inheritance.
